requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-experimental==0.0.47
langchain-openai==0.0.5
langchain-anthropic==0.1.1

# Vector Databases
chromadb==0.4.22
pinecone-client==2.2.4
qdrant-client==1.7.0

# Embeddings and Models
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
openai==1.6.1
anthropic==0.8.1

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.11.6
beautifulsoup4==4.12.2
lxml==4.9.3

# Search and Retrieval
rank-bm25==0.2.2
cohere==4.37

# API and Web Framework
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
pydantic==2.5.2
requests==2.31.0

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.25

# Monitoring and Evaluation
mlflow==2.8.1
ragas==0.1.6
langsmith==0.0.69

# Utilities
numpy==1.24.4
pandas==2.1.4
python-dotenv==1.0.0
tqdm==4.66.1
python-multipart==0.0.6

# Development and Testing
pytest==7.4.4
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Deployment
docker==7.0.0
kubernetes==28.1.0

